// src/app/claim-processing/manual-claim-processing/process-claim-nursing-charge/model/proc-clm-nursing-chrg-service.model.ts
import {MapLines1s} from './mapLines1s.model';
/**
 * Model class ProcClmNursingChrgService
 * Path: screenbean/procclmnursingchrgservice
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmnursingchrgservice::ProcClmNursingChrgService
 */
export class ProcClmNursingChrgService {
  mapLines1s: MapLines1s[] = [];
  m26cmnd = '';
  m26mem1 = '';
  m26mem2 = '';
  m26mem3 = '';
  m26cno1 = '';
  m26cno2 = '';
  m26cno3 = '';
  m26cno4 = '';
  m26cno5 = '';
  m26cno6 = '';
  m26com1 = '';
  m26com2 = '';
  m26name = '';
  m26pln1 = '';
  m26eff1 = '';
  m26trm1 = '';
  m26memn = '';
  m26assc = '';
  m26insr = '';
  m26pln2 = '';
  m26eff2 = '';
  m26trm2 = '';
  m26dob = '';
  m26pln3 = '';
  m26eff3 = '';
  m26trm3 = '';
  m26icd1 = '';
  m26icd2 = '';
  m26icd3 = '';
  m26pro1 = '';
  m26typ1 = '';
  m26sr11 = '';
  m26sr21 = '';
  m26nos1 = '';
  m26noh1 = '';
  m26chg1 = '';
  m26cpt1 = '';
  m26np1 = '';
  m26pl1 = '';
  m26pre1 = '';
  m26pro2 = '';
  m26typ2 = '';
  m26sr12 = '';
  m26sr22 = '';
  m26nos2 = '';
  m26noh2 = '';
  m26chg2 = '';
  m26cpt2 = '';
  m26np2 = '';
  m26pl2 = '';
  m26pre2 = '';
  m26pro3 = '';
  m26typ3 = '';
  m26sr13 = '';
  m26sr23 = '';
  m26nos3 = '';
  m26noh3 = '';
  m26chg3 = '';
  m26cpt3 = '';
  m26np3 = '';
  m26pl3 = '';
  m26pre3 = '';
  m26pro4 = '';
  m26typ4 = '';
  m26sr14 = '';
  m26sr24 = '';
  m26nos4 = '';
  m26noh4 = '';
  m26chg4 = '';
  m26cpt4 = '';
  m26np4 = '';
  m26pl4 = '';
  m26pre4 = '';
  m26pro5 = '';
  m26typ5 = '';
  m26sr15 = '';
  m26sr25 = '';
  m26nos5 = '';
  m26noh5 = '';
  m26chg5 = '';
  m26cpt5 = '';
  m26np5 = '';
  m26pl5 = '';
  m26pre5 = '';
  m26xpa1 = '';
  m26xpa2 = '';
  m26atrm = '';
  m26mpl1 = '';
  m26tos1 = '';
  m26str1 = '';
  m26end1 = '';
  m26shp1 = '';
  m26max1 = '';
  m26mpl2 = '';
  m26tos2 = '';
  m26str2 = '';
  m26end2 = '';
  m26shp2 = '';
  m26max2 = '';
  m26mpl3 = '';
  m26tos3 = '';
  m26str3 = '';
  m26end3 = '';
  m26shp3 = '';
  m26max3 = '';
  m26err1 = '';
  m26err2 = '';
  m26err3 = '';
}